# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "uiscloud_weknora 지식베이스 관리 시스템 API 문서",
        "title": "uiscloud_weknora API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "uiscloud_weknora Github",
            "url": "https://github.com/rockgis/uiscloud_weknora"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/sessions": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "현재 테넌트의 세션 목록을 조회합니다. 페이지네이션을 지원합니다",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "세션"
                ],
                "summary": "세션 목록 조회",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "페이지 번호",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "페이지당 항목 수",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "세션 목록",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "잘못된 요청 파라미터",
                        "schema": {
                            "$ref": "#/definitions/github_com_Tencent_uiscloud_weknora_internal_errors.AppError"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "새 대화 세션을 생성합니다",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "세션"
                ],
                "summary": "세션 생성",
                "parameters": [
                    {
                        "description": "세션 생성 요청",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler_session.CreateSessionRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "생성된 세션",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "잘못된 요청 파라미터",
                        "schema": {
                            "$ref": "#/definitions/github_com_Tencent_uiscloud_weknora_internal_errors.AppError"
                        }
                    }
                }
            }
        },
...
        "/sessions/{id}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "ID로 세션 상세 정보를 조회합니다",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "세션"
                ],
                "summary": "세션 상세 조회",
                "parameters": [
                    {
                        "type": "string",
                        "description": "세션 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "세션 상세 정보",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "세션이 존재하지 않습니다",
                        "schema": {
                            "$ref": "#/definitions/github_com_Tencent_uiscloud_weknora_internal_errors.AppError"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "세션 속성을 업데이트합니다",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "세션"
                ],
                "summary": "세션 업데이트",
                "parameters": [
                    {
                        "type": "string",
                        "description": "세션 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "세션 정보",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_Tencent_uiscloud_weknora_internal_types.Session"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "업데이트된 세션",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "세션이 존재하지 않습니다",
                        "schema": {
                            "$ref": "#/definitions/github_com_Tencent_uiscloud_weknora_internal_errors.AppError"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "지정된 세션을 삭제합니다",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "세션"
                ],
                "summary": "세션 삭제",
                "parameters": [
                    {
                        "type": "string",
                        "description": "세션 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "삭제 성공",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "세션이 존재하지 않습니다",
                        "schema": {
                            "$ref": "#/definitions/github_com_Tencent_uiscloud_weknora_internal_errors.AppError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_Tencent_uiscloud_weknora_internal_errors.AppError": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/github_com_Tencent_uiscloud_weknora_internal_errors.ErrorCode"
                },
                "details": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "github_com_Tencent_uiscloud_weknora_internal_errors.ErrorCode": {
            "type": "integer",
            "enum": [
                1000,
                1001,
                1002,
                1003,
                1004,
                1005,
                1006,
                1007,
                1008,
                1009,
                1010,
                2000,
                2001,
                2002,
                2003,
                2004,
                2100,
                2101,
                2102,
                2103
            ],
            "x-enum-varnames": [
                "ErrBadRequest",
                "ErrUnauthorized",
                "ErrForbidden",
                "ErrNotFound",
                "ErrMethodNotAllowed",
                "ErrConflict",
                "ErrTooManyRequests",
                "ErrInternalServer",
                "ErrServiceUnavailable",
                "ErrTimeout",
                "ErrValidation",
                "ErrTenantNotFound",
                "ErrTenantAlreadyExists",
                "ErrTenantInactive",
                "ErrTenantNameRequired",
                "ErrTenantInvalidStatus",
                "ErrAgentMissingThinkingModel",
                "ErrAgentMissingAllowedTools",
                "ErrAgentInvalidMaxIterations",
                "ErrAgentInvalidTemperature"
            ]
        },
...
        "github_com_Tencent_uiscloud_weknora_internal_types.ContextCompressionStrategy": {
            "type": "string",
            "enum": [
                "sliding_window",
                "smart"
            ],
            "x-enum-varnames": [
                "ContextCompressionSlidingWindow",
                "ContextCompressionSmart"
            ]
        },
        "github_com_Tencent_uiscloud_weknora_internal_types.ContextConfig": {
            "type": "object",
            "properties": {
                "compression_strategy": {
                    "description": "Compression strategy: \"sliding_window\" or \"smart\"",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_Tencent_uiscloud_weknora_internal_types.ContextCompressionStrategy"
                        }
                    ]
                },
                "max_tokens": {
                    "description": "Maximum tokens allowed in LLM context",
                    "type": "integer"
                },
                "recent_message_count": {
                    "description": "For sliding_window: number of messages to keep\nFor smart: number of recent messages to keep uncompressed",
                    "type": "integer"
                },
                "summarize_threshold": {
                    "description": "Summarize threshold: number of messages before summarization",
                    "type": "integer"
                }
            }
        },
...
        "github_com_Tencent_uiscloud_weknora_internal_types.FallbackStrategy": {
            "type": "string",
            "enum": [
                "fixed",
                "model"
            ],
            "x-enum-comments": {
                "FallbackStrategyFixed": "Fixed response",
                "FallbackStrategyModel": "Model fallback response"
            },
            "x-enum-descriptions": [
                "Fixed response",
                "Model fallback response"
            ],
            "x-enum-varnames": [
                "FallbackStrategyFixed",
                "FallbackStrategyModel"
            ]
        },
...
        "github_com_Tencent_uiscloud_weknora_internal_types.Session": {
            "type": "object",
            "properties": {
                "agent_config": {
                    "description": "에이전트 설정 (세션 수준, enabled와 knowledge_bases만 저장)",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_Tencent_uiscloud_weknora_internal_types.SessionAgentConfig"
                        }
                    ]
                },
                "context_config": {
                    "description": "컨텍스트 관리 설정 (선택 사항)",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_Tencent_uiscloud_weknora_internal_types.ContextConfig"
                        }
                    ]
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "description": {
                    "description": "Description",
                    "type": "string"
                },
                "embedding_top_k": {
                    "description": "벡터 검색 TopK",
                    "type": "integer"
                },
                "enable_rewrite": {
                    "description": "다중 턴 재작성 스위치",
                    "type": "boolean"
                },
                "fallback_response": {
                    "description": "고정 응답 내용",
                    "type": "string"
                },
                "fallback_strategy": {
                    "description": "폴백 전략",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_Tencent_uiscloud_weknora_internal_types.FallbackStrategy"
                        }
                    ]
                },
                "id": {
                    "description": "ID",
                    "type": "string"
                },
                "keyword_threshold": {
                    "description": "키워드 검색 임계값",
                    "type": "number"
                },
                "knowledge_base_id": {
                    "description": "Strategy configuration",
                    "type": "string"
                },
                "max_rounds": {
                    "description": "다중 턴 유지 횟수",
                    "type": "integer"
                },
                "rerank_model_id": {
                    "description": "재정렬 모델 ID",
                    "type": "string"
                },
                "rerank_threshold": {
                    "description": "재정렬 임계값",
                    "type": "number"
                },
                "rerank_top_k": {
                    "description": "재정렬 TopK",
                    "type": "integer"
                },
                "summary_model_id": {
                    "description": "요약 모델 ID",
                    "type": "string"
                },
                "summary_parameters": {
                    "description": "요약 모델 파라미터",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_Tencent_uiscloud_weknora_internal_types.SummaryConfig"
                        }
                    ]
                },
                "tenant_id": {
                    "description": "Tenant ID",
                    "type": "integer"
                },
                "title": {
                    "description": "Title",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "vector_threshold": {
                    "description": "벡터 검색 임계값",
                    "type": "number"
                }
            }
        },
        "github_com_Tencent_uiscloud_weknora_internal_types.SessionAgentConfig": {
            "type": "object",
            "properties": {
                "agent_mode_enabled": {
                    "description": "Whether agent mode is enabled for this session",
                    "type": "boolean"
                },
                "knowledge_bases": {
                    "description": "Accessible knowledge base IDs for this session",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "web_search_enabled": {
                    "description": "Whether web search is enabled for this session",
                    "type": "boolean"
                }
            }
        },
...
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
...
        "internal_handler_session.CreateSessionRequest": {
            "type": "object",
            "properties": {
                "agent_config": {
                    "description": "Agent configuration (optional, session-level config only: enabled and knowledge_bases)",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_Tencent_uiscloud_weknora_internal_types.SessionAgentConfig"
                        }
                    ]
                },
                "knowledge_base_id": {
                    "description": "ID of the associated knowledge base (optional, can be set/changed during queries)",
                    "type": "string"
                },
                "session_strategy": {
                    "description": "Session strategy configuration",
                    "allOf": [
                        {
                            "$ref": "#/definitions/internal_handler_session.SessionStrategy"
                        }
                    ]
                }
            }
        },
...
        "internal_handler_session.SessionStrategy": {
            "type": "object",
            "properties": {
                "embedding_top_k": {
                    "description": "Number of top results to retrieve from vector search",
                    "type": "integer"
                },
                "enable_rewrite": {
                    "description": "Whether to enable query rewrite for multi-round conversations",
                    "type": "boolean"
                },
                "fallback_response": {
                    "description": "Fixed response content for fallback",
                    "type": "string"
                },
                "fallback_strategy": {
                    "description": "Strategy to use when no relevant knowledge is found",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_Tencent_uiscloud_weknora_internal_types.FallbackStrategy"
                        }
                    ]
                },
                "keyword_threshold": {
                    "description": "Threshold for keyword-based retrieval",
                    "type": "number"
                },
                "max_rounds": {
                    "description": "Maximum number of conversation rounds to maintain",
                    "type": "integer"
                },
                "no_match_prefix": {
                    "description": "Prefix for responses when no match is found",
                    "type": "string"
                },
                "rerank_model_id": {
                    "description": "ID of the model used for reranking results",
                    "type": "string"
                },
                "rerank_threshold": {
                    "description": "Threshold for reranking results",
                    "type": "number"
                },
                "rerank_top_k": {
                    "description": "Number of top results after reranking",
                    "type": "integer"
                },
                "summary_model_id": {
                    "description": "ID of the model used for summarization",
                    "type": "string"
                },
                "summary_parameters": {
                    "description": "Parameters for the summary model",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_Tencent_uiscloud_weknora_internal_types.SummaryConfig"
                        }
                    ]
                },
                "vector_threshold": {
                    "description": "Threshold for vector-based retrieval",
                    "type": "number"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Bearer {token} 형식의 JWT 토큰을 입력하세요",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}